+++ git_well/main.py
@@ -17,6 +17,7 @@ class GitWellModalCLI(scfg.ModalCLI):
     from git_well.git_discover_remote import __cli__ as discover_remote
     from git_well.git_autoconf_gpgsign import __cli__ as autoconf_gpgsign
     from git_well.git_url_components import __cli__ as url
+    from git_well.patchdir.patchdir_modal import __cli__ as patchdir
 
 
 def main():