+++ git_well/patchdir/patchdir_modal.py
@@ -1,0 +1,13 @@
+import scriptconfig as scfg
+
+
+class GitWellPatchDirModalCLI(scfg.ModalCLI):
+    """
+    Subcommands for handling patch directories
+    """
+    __command__ = 'patchdir'
+    from git_well.patchdir.git_patchdir_save import __cli__ as save
+    from git_well.patchdir.git_patchdir_apply import __cli__ as apply
+
+
+__cli__ = GitWellPatchDirModalCLI