+++ CHANGELOG.md
@@ -6,6 +6,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.3.0 - Unreleased
 
+### Added
+
+* Added patchdir save / apply tools
+
 ### Removed
 * Drop Python 3.7 support